refactor: 히스토리 페이지 상태로직 훅으로 분리
