+++ front/src/pages/CrewHistory/index.tsx
@@ -5,11 +5,12 @@ import { AxiosError } from 'axios';
 import TableRow from '@components/TableRow';
 import EmptyContent from '@components/EmptyContent';
 import Filter from '@components/Filter';
+import useHistory from './hooks/useHistory';
 import { SnackbarContext } from '@context/SnackbarProvider';
 import { getCrewHistoriesByMe } from '@api/crew';
 import { cancelReservation } from '@api/reservation';
 import { ERROR_MESSAGE, ROUTES } from '@constants/index';
-import type { CrewHistory as CrewHistoryType, CrewHistoryStatus } from '@typings/domain';
+import type { CrewHistoryStatus } from '@typings/domain';
 import { logError } from '@utils/logError';
 
 import { theme } from '@styles/theme';
@@ -52,28 +53,8 @@ const historyStatus: HistoryStatus = {
 const CrewHistory = () => {
   const navigate = useNavigate();
   const showSnackbar = useContext(SnackbarContext);
-  const [historyList, setHistoryList] = useState<CrewHistoryType[]>([]);
+  const { historyList, setHistoryList, changeHistoryStatus, moveForefrontHistory } = useHistory();
   const [category, setCategory] = useState<string>('ALL');
-
-  const changeHistoryStatus = (reservationId: number, status: CrewHistoryStatus) => {
-    setHistoryList((prevHistory) => {
-      return prevHistory.map((history) => {
-        if (history.reservationId === reservationId) {
-          history.status = status;
-        }
-        return history;
-      });
-    });
-  };
-
-  const moveForefrontHistory = (reservationId: number) => {
-    const copyHistoryList = [...historyList];
-    const index = copyHistoryList.findIndex((history) => history.reservationId === reservationId);
-    const changedHistory = copyHistoryList[index];
-    copyHistoryList.splice(index, 1);
-    copyHistoryList.unshift(changedHistory);
-    setHistoryList(copyHistoryList);
-  };
 
   const handleShowSheet = (reservationId: number, status: string) => () => {
     navigate(`${ROUTES.CREW_SHEET}/${reservationId}`, { state: status });
